requirements.txt
Babel==2.2.0
colorlover==0.2.1
cufflinks==0.7.2
defusedxml==0.4.1
Django==1.9.4
django-allauth==0.25.2
django-bootstrap3==7.0.0
django-extra-views==0.6.4
django-guardian==1.4.2
django-haystack==2.4.1
django-json-field==0.5.7
django-oscar==1.2.1
django-oscar-accounts==0.3
django-polymorphic==0.9.1
django-tables2==1.0.7
django-treebeard==4.0
django-widget-tweaks==1.4.1
factory-boy==2.6.1
fake-factory==0.5.7
mock==1.3.0
mongoengine==0.10.6
numpy==1.11.0
oauthlib==1.0.3
pandas==0.18.0
pbr==1.8.1
phonenumbers==7.2.8
Pillow==2.9.0
plotly==1.9.6
psycopg2==2.6.1
purl==1.2
pymongo==3.2.1
python-dateutil==2.5.2
python3-openid==3.0.9
pytz==2015.7
requests==2.9.1
requests-oauthlib==0.6.1
simplejson==3.8.2
six==1.10.0
sorl-thumbnail==12.3
Unidecode==0.4.19
